requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.9.0
numba==0.59.1
numpy==1.22.0
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
six==1.16.0
